create job & delete job

+++ src/api/apiJobs.js
@@ -30,6 +30,8 @@ export const getJobs = async (token ,{location, company_id, searchQuery}) => {
 export const saveJob = async (token ,{isSavedJob}, savedData) => {
 
     const supabase = await supabaseClient(token);
+   console.log('is saved job:', isSavedJob);
+   console.log(' saved job data:', savedData);
    
     if(isSavedJob) {
 
@@ -37,6 +39,7 @@ export const saveJob = async (token ,{isSavedJob}, savedData) => {
             .from('saved_jobs')
             .delete()
             .eq('job_id', savedData.job_id)
+            .eq('user_id', savedData.user_id);
 
             if(deleteError) {
                 console.error('failed to delete saved sob' ,deleteError);
@@ -113,6 +116,57 @@ export async function addNewJob(token, _, jobData) {
 
     if(error) {
         console.log('Error creating Job:', error);
+        return null;
+    }
+    return data;
+}
+
+export async function getSavedJobs(token) {
+
+    
+    const supabase = await supabaseClient(token);
+   
+    const {data, error} = await supabase
+    .from("saved_jobs")
+    .select("*, job: jobs(*, company: companies(*))");
+
+    if(error) {
+        console.log('Error fetching saved Job:', error);
+        return null;
+    }
+    return data;
+}
+
+export async function getMyJobs(token, {recruiter_id}) {
+
+    
+    const supabase = await supabaseClient(token);
+   
+    const {data, error} = await supabase
+    .from("jobs")
+    .select("*, company: companies(*)")
+    .eq("recruiter_id", recruiter_id);
+
+    if(error) {
+        console.log('Error fetching my Jobs:', error);
+        return null;
+    }
+    return data;
+}
+
+export async function deleteJob(token, {job_id}) {
+
+    
+    const supabase = await supabaseClient(token);
+   
+    const {data, error: deleteError} = await supabase
+    .from("jobs")
+    .delete()
+    .eq("id", job_id)
+    .select();
+
+    if(deleteError) {
+        console.log('Error deleting Job:', deleteError);
         return null;
     }
     return data;
